Created delete dialog and added it to backlog item deletion, comment deletion, member deletion and project deletion

+++ corn-frontend/src/app/pages/boards/backlog/backlog-item-comment-list/backlog-item-comment/backlog-item-comment.component.ts
@@ -1,22 +1,24 @@
-import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
-import { BacklogItemComment } from "@interfaces/boards/backlog/backlog-item-comment";
-import { UserAvatarComponent } from "@pages/utils/user-avatar/user-avatar.component";
-import { MatError, MatFormField, MatHint, MatSuffix } from "@angular/material/form-field";
-import { CdkTextareaAutosize } from "@angular/cdk/text-field";
-import { MatInput } from "@angular/material/input";
-import { FormControl, ReactiveFormsModule, Validators } from "@angular/forms";
-import { CustomValidators } from "@core/validators/custom-validators";
-import { MatButton, MatIconButton } from "@angular/material/button";
-import { NgIcon, provideIcons } from "@ng-icons/core";
-import { matDelete, matDone, matEdit } from "@ng-icons/material-icons/baseline";
-import { NgIf } from "@angular/common";
-import { MatTooltip } from "@angular/material/tooltip";
-import { MatMenu, MatMenuItem, MatMenuTrigger } from "@angular/material/menu";
-import { MatIcon } from "@angular/material/icon";
+import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
+import {BacklogItemComment} from "@interfaces/boards/backlog/backlog-item-comment";
+import {UserAvatarComponent} from "@pages/utils/user-avatar/user-avatar.component";
+import {MatError, MatFormField, MatHint, MatSuffix} from "@angular/material/form-field";
+import {CdkTextareaAutosize} from "@angular/cdk/text-field";
+import {MatInput} from "@angular/material/input";
+import {FormControl, ReactiveFormsModule, Validators} from "@angular/forms";
+import {CustomValidators} from "@core/validators/custom-validators";
+import {MatButton, MatIconButton} from "@angular/material/button";
+import {NgIcon, provideIcons} from "@ng-icons/core";
+import {matDelete, matDone, matEdit} from "@ng-icons/material-icons/baseline";
+import {NgIf} from "@angular/common";
+import {MatTooltip} from "@angular/material/tooltip";
+import {MatMenu, MatMenuItem, MatMenuTrigger} from "@angular/material/menu";
+import {MatIcon} from "@angular/material/icon";
 import {
     BacklogItemCommentService
 } from "@core/services/boards/backlog/backlog-item-comment/backlog-item-comment.service";
-import { take } from "rxjs";
+import {take} from "rxjs";
+import {MatDialog} from "@angular/material/dialog";
+import {DeleteDialogComponent} from "@pages/utils/delete-dialog/delete-dialog.component";
 
 @Component({
     selector: 'app-backlog-item-comment',
@@ -42,7 +44,7 @@ import { take } from "rxjs";
     ],
     templateUrl: './backlog-item-comment.component.html',
     styleUrl: './backlog-item-comment.component.scss',
-    providers: [provideIcons({ matEdit, matDone, matDelete })],
+    providers: [provideIcons({matEdit, matDone, matDelete})],
 })
 export class BacklogItemCommentComponent implements OnInit {
 
@@ -54,12 +56,14 @@ export class BacklogItemCommentComponent implements OnInit {
 
     commentControl!: FormControl;
 
-    constructor(private commentService: BacklogItemCommentService) {}
+    constructor(private commentService: BacklogItemCommentService,
+                private dialog: MatDialog) {
+    }
 
     ngOnInit(): void {
         this.prepareComment();
 
-        this.commentControl =  new FormControl(this.comment.comment,
+        this.commentControl = new FormControl(this.comment.comment,
             [Validators.required, CustomValidators.notWhitespace()]);
     }
 
@@ -68,12 +72,12 @@ export class BacklogItemCommentComponent implements OnInit {
     }
 
     canEdit(): boolean {
-      //TODO acquire real information about user being owner of the comment or owner of the project
-      return true;
+        //TODO acquire real information about user being owner of the comment or owner of the project
+        return true;
     }
 
     editComment(): void {
-        if(this.commentControl.invalid) {
+        if (this.commentControl.invalid) {
             return;
         }
 
@@ -87,15 +91,26 @@ export class BacklogItemCommentComponent implements OnInit {
     }
 
     deleteComment(): void {
-        this.commentService.deleteComment(this.comment.backlogItemCommentId)
-            .pipe(take(1))
-            .subscribe(() => {
-                this.commentDeleted.emit();
-            });
+        const dialogRef = this.dialog.open(DeleteDialogComponent, {
+            enterAnimationDuration: '100ms',
+            exitAnimationDuration: '100ms'
+        })
+
+        dialogRef.afterClosed().pipe(take(1)).subscribe(result => {
+            if (!result) {
+                return;
+            }
+
+            this.commentService.deleteComment(this.comment.backlogItemCommentId)
+                .pipe(take(1))
+                .subscribe(() => {
+                    this.commentDeleted.emit();
+                });
+        })
     }
 
     formatDate(date: Date): string {
-        if(!date) {
+        if (!date) {
             return "";
         }
